+++ crates/matrix-sdk-crypto/src/session_manager/sessions.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use std::{
-    collections::{BTreeMap, BTreeSet, HashMap},
+    collections::{BTreeMap, BTreeSet},
     sync::{Arc, RwLock as StdRwLock},
     time::Duration,
 };
@@ -35,7 +35,7 @@ use crate::{
     error::OlmResult,
     gossiping::GossipMachine,
     requests::{OutgoingRequest, ToDeviceRequest},
-    store::{Changes, Result as StoreResult, Store, UserKeyQueryResult},
+    store::{Changes, Result as StoreResult, Store},
     types::{events::EventType, EventEncryptionAlgorithm},
     utilities::FailuresCache,
     ReadOnlyDevice,
@@ -63,14 +63,19 @@ pub(crate) struct SessionManager {
     wedged_devices: Arc<StdRwLock<BTreeMap<OwnedUserId, BTreeSet<OwnedDeviceId>>>>,
     key_request_machine: GossipMachine,
     outgoing_to_device_requests: Arc<StdRwLock<BTreeMap<OwnedTransactionId, OutgoingRequest>>>,
+
+    /// Servers that have previously appeared in the `failures` section of a
+    /// `/keys/claim` response.
+    ///
+    /// See also [`crate::identities::IdentityManager::failures`].
     failures: FailuresCache<OwnedServerName>,
+
     failed_devices: Arc<StdRwLock<BTreeMap<OwnedUserId, FailuresCache<OwnedDeviceId>>>>,
 }
 
 impl SessionManager {
     const KEY_CLAIM_TIMEOUT: Duration = Duration::from_secs(10);
     const UNWEDGING_INTERVAL: Duration = Duration::from_secs(60 * 60);
-    const KEYS_QUERY_WAIT_TIME: Duration = Duration::from_secs(5);
 
     pub fn new(
         users_for_key_claim: Arc<StdRwLock<BTreeMap<OwnedUserId, BTreeSet<OwnedDeviceId>>>>,
@@ -186,33 +191,6 @@ impl SessionManager {
         Ok(())
     }
 
-    async fn get_user_devices(
-        &self,
-        user_id: &UserId,
-    ) -> StoreResult<HashMap<OwnedDeviceId, ReadOnlyDevice>> {
-        use UserKeyQueryResult::*;
-
-        let user_devices = self.store.get_readonly_devices_filtered(user_id).await?;
-
-        let user_devices = if user_devices.is_empty() {
-            let cache = self.store.cache().await?;
-            match self
-                .key_request_machine
-                .identity_manager()
-                .key_query_manager
-                .wait_if_user_key_query_pending(cache, Self::KEYS_QUERY_WAIT_TIME, user_id)
-                .await?
-            {
-                WasPending => self.store.get_readonly_devices_filtered(user_id).await?,
-                _ => user_devices,
-            }
-        } else {
-            user_devices
-        };
-
-        Ok(user_devices)
-    }
-
     /// Get a key claiming request for the user/device pairs that we are
     /// missing Olm sessions for.
     ///
@@ -247,11 +225,17 @@ impl SessionManager {
         let mut missing: BTreeMap<_, BTreeMap<_, _>> = BTreeMap::new();
         let mut timed_out: BTreeMap<_, BTreeSet<_>> = BTreeMap::new();
 
-        // Add the list of devices that the user wishes to establish sessions
-        // right now.
-        for user_id in users.filter(|u| !self.failures.contains(u.server_name())) {
-            let user_devices = Box::pin(self.get_user_devices(user_id)).await?;
+        let unfailed_users = users.filter(|u| !self.failures.contains(u.server_name()));
 
+        // Get the current list of devices for each user.
+        let devices_by_user = Box::pin(
+            self.key_request_machine
+                .identity_manager()
+                .get_user_devices_for_encryption(unfailed_users),
+        )
+        .await?;
+
+        for (user_id, user_devices) in devices_by_user {
             for (device_id, device) in user_devices {
                 if !(device.supports_olm()) {
                     warn!(
@@ -270,7 +254,7 @@ impl SessionManager {
                         true
                     };
 
-                    let is_timed_out = self.is_user_timed_out(user_id, &device_id);
+                    let is_timed_out = self.is_user_timed_out(&user_id, &device_id);
 
                     if is_missing && is_timed_out {
                         timed_out.entry(user_id.to_owned()).or_default().insert(device_id);
@@ -571,6 +555,7 @@ mod tests {
         iter,
         ops::Deref,
         sync::{Arc, RwLock as StdRwLock},
+        time::Duration,
     };
 
     use matrix_sdk_test::{async_test, response_from_file};
@@ -582,7 +567,7 @@ mod tests {
             },
             IncomingResponse,
         },
-        device_id, owned_server_name, user_id, DeviceId, UserId,
+        device_id, owned_server_name, user_id, DeviceId, OwnedUserId, UserId,
     };
     use serde_json::json;
     use tokio::sync::Mutex;
@@ -761,6 +746,45 @@ mod tests {
         info!("Key claim request: {:?}", keys_claim_request.one_time_keys);
         let bob_key_claims = keys_claim_request.one_time_keys.get(bob.user_id()).unwrap();
         assert!(bob_key_claims.contains_key(bob_device.device_id()));
+    }
+
+    #[async_test]
+    async fn test_session_creation_does_not_wait_for_keys_query_on_failed_server() {
+        let (manager, identity_manager) = session_manager_test_helper().await;
+
+        // We start tracking Bob's devices.
+        let other_user_id = OwnedUserId::try_from("@bob:example.com").unwrap();
+        {
+            let cache = manager.store.cache().await.unwrap();
+            identity_manager
+                .key_query_manager
+                .synced(&cache)
+                .await
+                .unwrap()
+                .update_tracked_users(iter::once(other_user_id.as_ref()))
+                .await
+                .unwrap();
+        }
+
+        // Do a keys query request, in which Bob's server is a failure.
+        let (key_query_txn_id, _key_query_request) =
+            identity_manager.users_for_key_query().await.unwrap().pop_first().unwrap();
+        let response = KeysQueryResponse::try_from_http_response(response_from_file(
+                &json!({ "device_keys": {}, "failures": { other_user_id.server_name(): "unreachable" }})
+        )).unwrap();
+        identity_manager.receive_keys_query_response(&key_query_txn_id, &response).await.unwrap();
+
+        // Now, an attempt to get the missing sessions should now *not* block. We use a
+        // timeout so that we can detect the call blocking.
+        let result = tokio::time::timeout(
+            Duration::from_millis(10),
+            manager.get_missing_sessions(iter::once(other_user_id.as_ref())),
+        )
+        .await
+        .expect("get_missing_sessions blocked rather than completing quickly")
+        .expect("get_missing_sessions returned an error");
+
+        assert!(result.is_none(), "get_missing_sessions returned Some(...)");
     }
 
     // This test doesn't run on macos because we're modifying the session